rename wait-for to wait
